--- storage_image/views/storage_image_wizard.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
-    <record id="storage_image_view_wizard" model="ir.ui.view">
-        <field name="model">storage.image.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Upload an image">
-                <group name="main">
-                    <field name="the_file" filename="name"/>
-                    <field name="name" />
-                    <field name="alt_name" />
-                </group>
-                <footer>
-                    <button name="associate" string="Accociate"
-                        type="object" class="oe_highlight"/>
-                    <button special="cancel" string="Cancel" class="oe_link"/>
-                </footer>
-           </form>
-        </field>
-    </record>
-
-    <record id="action_wiz_associate" model="ir.actions.act_window">
-        <field name="name">Upload an image</field>
-        <field name="res_model">storage.image.wizard</field>
-        <field name="view_type">form</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-    </record>
-</data>
-</openerp>
